@@ -9,23 +9,21 @@
   },
   "dependencies": {
     "firebase": "^9.2.0",
-    "vue": "^2.6.11",
-    "vue-router": "^3.5.2",
+    "vue": "^3.1.0",
+    "vue-router": "^4.3.0",
     "vuedraggable": "^2.24.3",
-    "vuetify": "^2.4.0",
-    "vuex": "^3.6.2"
+    "vuetify": "^3.5.6",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue2": "^2.3.1",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vue/compiler-sfc": "^3.1.0",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.2.2",
     "sass": "~1.32.0",
     "sass-loader": "^10.0.0",
     "unplugin-vue-components": "^0.26.0",
-    "vite": "^5.1.4",
-    "vue-cli-plugin-vuetify": "~2.4.3",
-    "vue-template-compiler": "^2.6.11",
-    "vuetify-loader": "^1.7.0"
+    "vite": "^5.1.4"
   },
   "eslintConfig": {
     "root": true,
