@@ -48,7 +48,7 @@ export default {
       listImage: ListImage
     }
   },
-  beforeDestroy() {
+  beforeUnmount() {
     this.$store.commit("setActiveBoard", this.getBoard);
     this.$store.commit("setActivePage", "default");
   },
