@@ -1,15 +1,14 @@
-import Vue from "vue";
-import Vuex from "vuex";
-
+import { createStore as _createStore } from 'vuex';
 import state from "./state";
 import actions from "./actions";
 import mutations from "./mutations";
 import getters from "./getters";
 
-Vue.use(Vuex);
-export default new Vuex.Store({
-  state,
-  mutations,
-  actions,
-  getters,
-});
+export function createStore(){
+  return _createStore({
+    state,
+    mutations,
+    actions,
+    getters,
+  })
+}
